replaced floats with doubles

+++ src/NetworkInstances/LinearRegression.cxx
@@ -16,23 +16,23 @@ class LinearRegressionLayer : public NeuralNetwork::Layer
             setParameter(1, 0);
         }
     private:
-        void op(const std::vector<float> &input,
-                std::vector<float> &output) const override
+        void op(const std::vector<double> &input,
+                std::vector<double> &output) const override
         {
-            float a = getParameter(0);
-            float b = getParameter(1);
-            float x = input[0];
+            double a = getParameter(0);
+            double b = getParameter(1);
+            double x = input[0];
 
             output[0] = a * x + b;
         }
-        void bprop(const std::vector<float> &input,
-                   const std::vector<float> &outputGradient,
-                   std::vector<float> &inputGradient,
-                   std::vector<float> &parameterGradient) const override
+        void bprop(const std::vector<double> &input,
+                   const std::vector<double> &outputGradient,
+                   std::vector<double> &inputGradient,
+                   std::vector<double> &parameterGradient) const override
         {
-            float a = getParameter(0);
-            //float b = getParameter(1);
-            float x = input[0];
+            double a = getParameter(0);
+            //double b = getParameter(1);
+            double x = input[0];
 
             // d(out)/d(param)
             parameterGradient[0] = x * outputGradient[0];
@@ -51,19 +51,19 @@ class MeanSquaredErrorLayer : public NeuralNetwork::ErrorLayer
         {
         }
     private:
-        void op(const std::vector<float> &input, std::vector<float> &output) const override
+        void op(const std::vector<double> &input, std::vector<double> &output) const override
         {
             const auto &expectedInput = getExpectedResult();
 
-            float sum = 0.0f;
+            double sum = 0.0f;
             for (size_t i = 0; i < input.size(); i++)
             {
-                float difference = input[i] - expectedInput[i];
+                double difference = input[i] - expectedInput[i];
                 sum += difference * difference;
             }
             output[0] = sum;
         }
-        void bprop(const std::vector<float> &input, const std::vector<float> &outputGradient, std::vector<float> &inputGradient, std::vector<float> &) const override
+        void bprop(const std::vector<double> &input, const std::vector<double> &outputGradient, std::vector<double> &inputGradient, std::vector<double> &) const override
         {
             const auto &expectedInput = getExpectedResult();
 
@@ -89,10 +89,10 @@ LinearRegression::LinearRegression()
     std::mt19937 rng(std::chrono::system_clock::now().time_since_epoch().count());
     for (int i = 0; i < 1000; i++)
     {
-        std::normal_distribution<float> gauss{0, 0.2};
-        std::uniform_real_distribution<float> uniform{0, 5};
-        float x = uniform(rng);
-        float y = (3 - 0.8 * x) + gauss(rng);
+        std::normal_distribution<double> gauss{0, 0.2};
+        std::uniform_real_distribution<double> uniform{0, 5};
+        double x = uniform(rng);
+        double y = (3 - 0.8 * x) + gauss(rng);
         //if (x > 2.5)
         //    y = (3 - 0.8 * (5 - x)) + gauss(rng);
 
@@ -102,30 +102,30 @@ LinearRegression::LinearRegression()
 
 
 
-void LinearRegression::addTrainingSample(float x, float y)
+void LinearRegression::addTrainingSample(double x, double y)
 {
-    std::vector<float> in = {x};
-    std::vector<float> out = {y};
+    std::vector<double> in = {x};
+    std::vector<double> out = {y};
     network->addTrainingSample(in, out);
 }
 
 void LinearRegression::run()
 {
-    float error = 1000;
+    double error = 1000;
     while (true)
     {
-        //std::cout << "Current value at 1: " << network->compute(std::vector<float>({1})).at(0) << std::endl;
-        //std::cout << "Current value at 3: " << network->compute(std::vector<float>({3})).at(0) << std::endl;
-        //std::cout << "Current value at 4: " << network->compute(std::vector<float>({4})).at(0) << std::endl;
-        float previousError = network->trainingStep(0.03);
+        //std::cout << "Current value at 1: " << network->compute(std::vector<double>({1})).at(0) << std::endl;
+        //std::cout << "Current value at 3: " << network->compute(std::vector<double>({3})).at(0) << std::endl;
+        //std::cout << "Current value at 4: " << network->compute(std::vector<double>({4})).at(0) << std::endl;
+        double previousError = network->trainingStep(0.03);
         if (std::abs(previousError - error) < 0.00000001)
             break;
         error = previousError;
         std::cout << "Error before trainingStep: " << previousError << std::endl;
         //std::this_thread::sleep_for(std::chrono::milliseconds(100));
     }
-    float a = network->getParameter(0, 0);
-    float b = network->getParameter(0, 1);
+    double a = network->getParameter(0, 0);
+    double b = network->getParameter(0, 1);
     std::cout << "Result: " << a << " * x " << (b < 0 ? "- " : "+ ") << std::abs(b) << std::endl;
 }
 
@@ -133,11 +133,11 @@ void LinearRegression::run()
 namespace
 {
 
-const float MIN_X = -1.0f;
-const float MAX_X = 6.0f;
-const float MIN_Y = -2.0f;
-const float MAX_Y = 5.0f;
-void convertAddressToPixel(SDL_Rect &rect, float x, float y, int &px, int &py)
+const double MIN_X = -1.0f;
+const double MAX_X = 6.0f;
+const double MIN_Y = -2.0f;
+const double MAX_Y = 5.0f;
+void convertAddressToPixel(SDL_Rect &rect, double x, double y, int &px, int &py)
 {
 
     px = int((x - MIN_X) / (MAX_X - MIN_X) * rect.w + 0.5f);
@@ -158,8 +158,8 @@ void LinearRegression::render(SDL_Renderer *renderer, SDL_Rect &rect)
     {
         int px, py;
 
-        float pointX = network->getTrainingInput(i)[0];
-        float pointY = network->getTrainingOutput(i)[0];
+        double pointX = network->getTrainingInput(i)[0];
+        double pointY = network->getTrainingOutput(i)[0];
 
         convertAddressToPixel(rect, pointX, pointY, px, py);
 
@@ -169,8 +169,8 @@ void LinearRegression::render(SDL_Renderer *renderer, SDL_Rect &rect)
     }
 
     {
-        float a = network->getParameter(0, 0);
-        float b = network->getParameter(0, 1);
+        double a = network->getParameter(0, 0);
+        double b = network->getParameter(0, 1);
         int x0, y0, x1, y1;
         convertAddressToPixel(rect, MIN_X, a * MIN_X + b, x0, y0);
         convertAddressToPixel(rect, MAX_X, a * MAX_X + b, x1, y1);
