+++ src/NetworkInstances/LinearRegression.hxx
@@ -11,7 +11,7 @@ class LinearRegression : public SDL2Subwindow
 {
     public:
         LinearRegression();
-        void addTrainingSample(float x, float y);
+        void addTrainingSample(double x, double y);
         void run();
 
         void render(SDL_Renderer *renderer, SDL_Rect &rect) override;
